# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Flood Risk API
  version: 0.1.0-draft
  summary: API contract draft for the Flood Risk public and admin APIs.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: >
    Draft contract derived from PROJECT_SDD sections 7 Domain Model,
    9 Risk Scoring Design, and 10 API Contract Draft. Public risk levels
    are the Traditional Chinese labels `低`, `中`, `高`, `極高`, and `未知`.
    Query heat is exposed only as public attention metadata and must not be
    used as a realtime or historical risk score input.
servers:
  - url: /
    description: Same-origin deployment
security: []
...
paths:
...
  /admin/v1/jobs:
    get:
      tags:
        - Admin
      operationId: listAdminJobs
      summary: List background job status.
      security:
        - AdminBearerAuth: []
      x-required-roles:
        - admin
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/JobStatus"
        - name: job_key
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 120
      responses:
        "200":
          description: Job status rows.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminJobsResponse"
              examples:
                jobs:
                  value:
                    jobs:
                      - job_key: ingest.cwa.rainfall
                        adapter_key: cwa.rainfall
                        started_at: "2026-04-28T12:50:00Z"
                        finished_at: "2026-04-28T12:55:00Z"
                        status: succeeded
                        items_fetched: 120
                        items_promoted: 118
                        items_rejected: 2
                        error_code: null
                        error_message: null
                        source_timestamp_min: "2026-04-28T12:00:00Z"
                        source_timestamp_max: "2026-04-28T12:50:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /admin/v1/sources:
    get:
      tags:
        - Admin
      operationId: listAdminSources
      summary: List source health and data freshness.
      security:
        - AdminBearerAuth: []
      x-required-roles:
        - admin
      parameters:
        - name: health_status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/HealthStatus"
      responses:
        "200":
          description: Data source health and freshness metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSourcesResponse"
              examples:
                sources:
                  value:
                    sources:
                      - id: cwa-rainfall
                        name: "中央氣象署雨量資料"
                        adapter_key: cwa.rainfall
                        source_type: official
                        license: "Open Government Data License"
                        update_frequency: PT10M
                        last_success_at: "2026-04-28T12:55:00Z"
                        last_failure_at: null
                        health_status: healthy
                        legal_basis: L1
                        source_timestamp_min: "2026-04-28T12:00:00Z"
                        source_timestamp_max: "2026-04-28T12:50:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Opaque admin token
      description: >
        Admin API bearer token. The Phase skeleton accepts a configured opaque
        admin token and treats it as the `admin` role. Production auth may move
        to JWT or another RBAC provider, but every operation tagged Admin must
        continue to require an authenticated admin principal.
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Unauthorized:
      description: Missing, expired, or invalid admin bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated principal does not have the required admin role.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    SourceType:
      type: string
      enum:
        - official
        - news
        - forum
        - social
        - user_report
        - derived
...
    AdminJobsResponse:
      type: object
      required:
        - jobs
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/IngestionJob"
      additionalProperties: false
    IngestionJob:
      type: object
      required:
        - job_key
        - status
        - items_fetched
        - items_promoted
        - items_rejected
      properties:
        job_key:
          type: string
        adapter_key:
          type:
            - string
            - "null"
        started_at:
          type:
            - string
            - "null"
          format: date-time
        finished_at:
          type:
            - string
            - "null"
          format: date-time
        status:
          $ref: "#/components/schemas/JobStatus"
        items_fetched:
          type: integer
          minimum: 0
        items_promoted:
          type: integer
          minimum: 0
        items_rejected:
          type: integer
          minimum: 0
        error_code:
          type:
            - string
            - "null"
        error_message:
          type:
            - string
            - "null"
        source_timestamp_min:
          type:
            - string
            - "null"
          format: date-time
        source_timestamp_max:
          type:
            - string
            - "null"
          format: date-time
      additionalProperties: false
    JobStatus:
      type: string
      enum:
        - queued
        - running
        - succeeded
        - failed
        - skipped
        - disabled
    AdminSourcesResponse:
      type: object
      required:
        - sources
      properties:
        sources:
          type: array
          items:
            $ref: "#/components/schemas/DataSource"
      additionalProperties: false
    DataSource:
      type: object
      required:
        - id
        - name
        - adapter_key
        - source_type
        - license
        - update_frequency
        - health_status
        - legal_basis
      properties:
        id:
          type: string
        name:
          type: string
        adapter_key:
          type: string
        source_type:
          $ref: "#/components/schemas/SourceType"
        license:
          type: string
        update_frequency:
          type: string
          description: ISO 8601 duration or human-readable source cadence.
        last_success_at:
          type:
            - string
            - "null"
          format: date-time
        last_failure_at:
          type:
            - string
            - "null"
          format: date-time
        health_status:
          $ref: "#/components/schemas/HealthStatus"
        legal_basis:
          $ref: "#/components/schemas/LegalBasis"
        source_timestamp_min:
          type:
            - string
            - "null"
          format: date-time
        source_timestamp_max:
          type:
            - string
            - "null"
          format: date-time
      additionalProperties: false
    HealthStatus:
      type: string
      enum:
        - healthy
        - degraded
        - failed
        - disabled
        - unknown
    LegalBasis:
      type: string
      description: Data legality level from the SDD source policy.
      enum:
        - L1
        - L2
        - L3
        - L4
        - L5
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type:
                - object
                - array
                - string
                - "null"
            request_id:
              type:
                - string
                - "null"
          additionalProperties: false
      additionalProperties: false